Add files via upload
An amazing exploration of the mind is now possible for everyone. With the Colors of The Sunrise, the first volume of the series The Psychotherapy of Whole: Aesthetics, Philosophy, Humanism, and Cognitive Science the reader has the opportunity to engage with a book that utilizes the methods and structure of self-help, popular science, and expressive therapies books, to change all the sides of the human experience as it manifests in the social, psychological, bodily, and cognitive life.

Science, psychotherapy, philosophy, music, art and digital reality for the first time come together in a book phenomenon and a series designed during 16 years to provide the first A.I Psychotherapy model internationally, focused on a profound study that has been evaluated by leading names from many of the areas analyzed around the world.

The reader gradually introduced to a healing result as it is offered clinical treatment to the senses, the psychological reality, and the cognitive mind structures, activating inactive cerebrum areas and getting informed about the manner of changing their behavior and thought. The direct healing and the way to acquire therapeutic results in everyday life are supported by explanations that become easily understood by everyone. The book has been based on the background of advanced academic research, lending from the recent and updated investigations in an extraordinary number of areas. These include but are not limited to the disciplines of social sciences, psychiatry, philosophy, expressive art therapies, exact sciences, history, politics, artificial intelligence, and humanities that are presented from a global perspective.

It is an exceptional interpretation of human cognitive structures that satisfy the curiosity and assist everyone in developing their intelligence, and investigate their self, acquiring a balanced theory of mind that can be applied removing the barriers of the daily dichotomy, which everyone encounters while conceptualizing reality. Linguistics, neuroscience, anthropology, robotics, physics, and mathematics reconsidered as positive parts of the universal structures. It is explained the manner of uniting the cognition staying in perfect harmony with the today’s knowledge acquired by cognitive science, and the knowledge that traditionally contributes to the flourishing of the human mind through the philosophical, and traditional approaches, as they have been described during the centuries globally.

The book’s design supports sound during the reading and acquaintance with the scientific methodology providing the chance to the reader to benefit from the purpose of the book that is open to the dialogue and facilitates the critical thought. It is utilized the state of the art tool of artificial intelligence reaching to an interdisciplinary approach of logic and therapy of cognitive mind structures based on the reunion of knowledge and logic as it manifests in human mind universally.

Through the familiarization with the approach, the reader experiences and learns a functional manner of relating to their body and environment, changing their dysfunctional behavior concerned with many split cognitive structures.

This manner discovered through the cutting-edge clinical and scientific study of the author. As a Licensed Social Worker and Cognitive Scientist, Anthony N. Skarvelakis devoted many years of professional investigation to reach a cognitive metaprogram as a model that expands human thought, giving the solution searched for years regarding the design of the first complete model of A.I Psychotherapy globally. Τhis program applies successfully to human and artificial intelligence. It is a program, (using a term employed in academia with the same meaning in A.I, robotics, cognitive science, cognitive psychology, cognitive oriented psychotherapeutic approaches), of critical importance today that lays at the intersection of research in Artificial Intelligence and Psychotherapy.